fixed password merge
